Fix broken text selection
After aosp/1562116 we had a broken text selection due to the incorrect x/y positions we get from the system when the view is in a Popup and a Popup has been just moved. I fixed it currently with not using the motionEvent.getX() and instead calculating the position in the view manually from the motionEvent.getRawX/Y().

Bug: 180029223
Test: manually, filed a bug b/180029223 to find the final solution and cover with a proper integration tests
Change-Id: I9128485838426a8c116c49338049515e4c1b2444