Reshuffling protos to allow them to be included from two libraries.
Previously, we shaded wear-tiles-proto into wear-tiles, and renamed
com.google.protobuf to androidx.wear.tiles.protobuf. We need to do the
same for the renderer (potentially with a different rename), but if we
do that, an application which included both libs can't compile (as the
generated proto classes also clash).

Instead, this refactors the generated protos, and the shaded libprotobuf
into wear-tiles-proto, which is then included from either wear-tiles or
wear-tiles-renderer. The API for this is currently empty. This is a bug;
we should be tracking that private API.

There's some hackery to make all this build properly, but I can build
both libs into a test app correctly now.

Test: Currently manual. Existing tests pass.
Relnote: N/A
Change-Id: I75cd9cacc83002246b507d5a1be55973273aef0e