Only compare accesibility action label in semantics tree diffing
During recompose, even if a semantics node is unchanged, and if
this node has an accessibility action, semantics tree diffing
will think the accessibility action changed and send out a window
content change event. This caused snackbar to announce when it
disappears if we don't remove the live region upon disappearing.

According to discussion with pweaver, accessibility action callback
is not reported to accessibility services(only the action id and
label are reported), so we can compare the action label only. This
will help to remove the snackbar workaround in aosp/1580798.

Test: unit tests added.

Change-Id: I6d9dbd729049e219d4973a1ad857ca3c2e3f21c9