Keep any active launchers in the ActivityResultRegistry
After an ActivityResultLauncher has launched an Activity, it is
currently possible for the launching Activity to be destroyed, which
would result in any activity results being lost.

Instead of allowing there active launcher to be dropped, we should
maintain them, ensuring that if an ActivityResult is launched, it always
gets a result.

RelNote: "ActivityResultLaunchers that have launched, will now always
return their results."
Test: ActivityResultRegistryTest
Bug: 181267562

Change-Id: I9ef34256a1e0f93fd1fd8983e867b9dd5e8e2441