Move more benchmarks.
This moves the rest of the benchmarks to their separate modules. Except
lazy that is currently being worked on so needs to stay. I have also
introduced compose-benchmark-utils to avoid general tests depending on
benchmark library. I did not change the package in this CL though as it
would modify lots of files and make this CL hard to read.

Bug: 179734581
Bug: 181687019
Test: Benchmarks run
Change-Id: Ifbfde52abb8e70134395134a9848753a851a42d0