Fix MutableLiveData lint rule false report
If using multiple MutableLiveData variables as fields and mixed the
nullability, the lint rule could falsely report that a livedata value
was improperly being made null. This is because as constructed, it
didn't match up the fields with the assignments.

This change maps the field variable names to the reference types to
ensure that when reporting, the rule knows how to match them.

This also cleans up the rule by changing it to using a UastScanner.

Relnote: "The `NonNullableMutableLiveData` lint rule can now properly
differentiate between field variables with different nullability."
Test: added nullLiterallFailMultipleFieldsDifferent
Bug: 169249668

Change-Id: Ieb0270de1b22f3a10641643438bc2b9f1f5d72e8