Cleanup after change: 903e9d1
In change: 903e9d1 CornerSize was handled by adding the
InspectableValue interface to all the private implementations of
CornerSize.

In this change we are reenabling testCornerBasedShape.
This test revealed an undesired effect of the current code:

A value believed to be a composite value which turns out NOT to
have child elements should NOT be tracked as a value we may later
create a reference to. This bug has been fixed.

Bug: 180952854
Test: ParameterFactoryTest

31b7c7d1d7ded50470aadd594d73116c4

Change-Id: I3ff93f13856c6af0912bca21b734ee26cf6236d9