Make LazyPagingItems' itemCount and item getter observable
This change removes hack recomposerPlaceholder we had to trigger the recompositions. It allows LazyPagingItems usage outside of official item/itemIndexed functions, for example with experimental LazyVerticalGrid.
Also this should help with recompositions granularity. For example if we receive an event from Pager that the item 2 has been changed we can only recompose this item.

Relnote: LazyPagingItems' itemCount and item getter are now observable which allows it to be used with LazyVerticalGrid as well
Test: new tests in LazyPagingItemsTest
Bug: 171872064
Bug: 168285687
Change-Id: Ie24468ec51660c144acab71e8e520ac09d00b023