identity: Fix buggy parsing of (r, s) in COSE_Sign1 causing flaky tests.
ProvisioningTest has been observed to be flaky and I was able to
create a minimal test case pin-pointing it to a bug in how ECDSA
signatures were converted between DER and COSE_Sign1 encoding.
The problem was that we converted the bytes of r and s to an
signed integer when we should have been using unsigned integers
instead.

In particular UtilUnitTests#coseSignAndVerify would fail a couple
of times every hundred runs. After fixing it, ran it for 10000 times
without failure.

Bug: 180309735
Test: Unit tests are no longer flaky
Change-Id: I8972d9d68f5cb4367ca3c58982aa62c69347d718