Fix NPE when using ActivityResultContract with non-null Unit input
When using a custom ActivityResultContract that has a non-null Unit
input, you can use the launch() extension methond that requires no
parameters. In this case, that call currently fail because we internal
call launch() with `null` instead of `Unit`.

RelNote: "Using the `launch()` method extension with a custom
`ActivityResultContract` that has a `Unit` input will no longer cause a
`NullPointerException`"
Test: ActivityResultLauncherTest
Bug: 183837954

Change-Id: I76282e17d47d25e3ff859f0b646e6063a381a510