[GH] [Navigation] Fix deeplink domain parsing being case sensitive
## Proposed Changes

The current regex pattern matcher was doing a case sensitive check of the domain and arguments. Simply changing it to case insensitive seems to be all that's required to fix this bug. It is important that parameters are still case sensitive but this is still the case as parameter matching is done by the `Uri` code not the `NavDeepLink` code. I have written tests to cover these cases.

If it is deemed too risky to just run the pattern matcher with `CASE_INSENSITIVE` I believe the regex will have to be redesigned not to use `\Q\E`.

## Testing

- Tested using the sample project and deep linking via `adb shell am start -a android.intent.action.VIEW -d https://www.Iana.org/domains/second`
- Added new unit tests to the project for the different case insensitive and sensitive sections

Test: /gradlew test connectedCheck

## Issues Fixed
https://issuetracker.google.com/issues/153829033
Fixes: b/153829033

This is an imported pull request from https://github.com/androidx/androidx/pull/144.

Resolves #144
Github-Pr-Head-Sha: 868bcf3346af6eb46da7e5988977574be8843eba
GitOrigin-RevId: 0c507315877618e865f6a79ce4e01ca594da28a9
Change-Id: Ic03b432ffc9f8a71fccad3f01c13232561366d46