Keep RecordingInputConnection.textFieldValue up to date
When a change in TextFieldValue was carried to
TextInputServiceAndroid, we had a shortcut that early
returned if the old value is the same as new value.

However, this caused RecordingInputConnection.textFieldValue
to become obsolete.

This caused issues with functions in RecordingInputConnection
that accessed the TextFieldValue.

This CL makes sure that RecordingInputConnection.textFieldValue
is always up to date.

Test: Manual test on Pixel 4, Android 11, SwiftKey
Test: Manual test on Pixel 4, Android 11, Gboard
Test: Manual test on Samsung S8, Android 9, Samsung keyboard
Test: Manual test on Samsung S8, Android 9, SwiftKey

Test: ./gradlew compose:ui:ui:test
Test: ./gradlew compose:ui:ui:connectedAndroidTest

Bug: 177236401
Bug: 177732712
Bug: 183851532
Change-Id: Ied6195456aa4a6b9f3d1d4e98cd694b10ff31a76