Save ViewModels when saving and restore the back stack
The most important part of calling saveBackStack() is
that the state of the fragments must be saved.

This can be roughly summarized in three parts:
1) The Fragment's saved instance state
2) The Fragment's View state
3) The Fragment's non-config state

The first two are already handled by the previous
changes, but the later one is just as important,
particularly when it comes to immediately showing
cached data. As retained fragments are not
supported when using saveBackStack(), the only
non-config state is that of ViewModels.

Besides just avoiding clearing the ViewModels when
the fragment is destroyed as part of saving the
back stack, special attention was taken to the
case where the Activity / parent Fragment is itself
destroyed. In this case, any ViewModels associated
with saved fragments must be specifically cleared as
they would not automatically be cleaned up on their
own (as their containing fragment no longer exists,
just its saved state).

Test: updated tests pass
BUG: 80029773
Change-Id: Idafc0af58e1a43893774f6ae6d7caac8bd56efb4