Fix accessibility scroll actions
Scroll forward/backward/left/up/right/down are supposed to scroll by one
screen, but they were scrolling to the end of the scrollable (in the
right direction though).

Fix uses the viewport to determine the amount by which to scroll, and
adds tests for these actions when applied to horizontal/verticalScroll.

Fix: 184153461
Test: ./gradlew compose:foundation:foundation:cC
Relnote: "Fixed `ACTION_SCROLL_FORWARD`, `ACTION_SCROLL_BACKWARD`,
`accessibilityActionScrollLeft`, `accessibilityActionScrollUp`,
`accessibilityActionScrollRight` and `accessibilityActionScrollDown`
accessibility scroll actions. Instead of scrolling to the end of the
scrollable, it will now scroll by one screen in the given direction."

Change-Id: Ieccb0a794d61dc5fe28a236b379755d776c023dc