AndroidX Webkit: fix ordering issues on AssetLoader.Builder
We observed some WebViewAssetLoader.Builder methods are order-dependent
in a surprising way. This fixes the builder to no longer care about the
ordering of setDomain and setHttpAllowed in relation to addPathHandler
calls. This adds two tests for this case.

There are still good use cases for addPathHandler() calls to be
order-dependent. I noticed this wasn't actually covered by any
integration tests, so I added one to cover this case as well. I updated
addPathHandler's javadoc to clarify that order matters, although this
was already documented on the PathHandler interface javadoc.

Fixes: 182196765
Test: ./gradlew :webkit:webkit:connectedAndroidTest
Relnote: "Fixed a bug where WebViewAssetLoader.Builder methods were
    unintentionally order-dependent."
Change-Id: If420d559a21ea624b2a493d09550ed496ffee392