Prevent redundant ListUpdateCallback events from firing
Avoid sending Differ events with count == 0 where they would essentially
be no-ops.

This conflicts with using RecyclerView.AdapterDataObserver to set state
restoration policy as it causes it to fail to set in the empty case; in
order to continue supporting this it has been moved to depend on
LoadState (which is a bit roundabout, but should be equivalent).

Relnote: "Fixed an issue where Paging would sometimes send no-op differ
  events to RecyclerView, which could cause certain listeners to trigger
  early."
Bug: 182510751
Test: ./gradlew paging:paging-runtime:cC
Change-Id: Ic507fa793cb22ac9f8880a8f8841f8e6caa197f4