Export provider info as part of editor session.
This commit adds information about the provider of a complication as
part of the editor session. This means that the watch face can get
access to this information in the editor without having to invoke the
ProviderInfoRetriver directly and we can process the update coming
from the provider configuration flow directly.

To be able to test this end-to-end, this changelist adds a tooltip to
the configuration activity such that long-pressing on the complication
to be configured shows the name of the complication provider currently
associated with it. Since tooltips are only supported on API level 26
and above, this behavior is not added on earlier versions of the OS.

Bug: 183910046
Test: tested with sample watch face
Relnote: "We added information about the complication provider for each complication within EditorSession."
Change-Id: I37f14f4867c495bfd9b67d9e32318e5a78eb33c7