Run QueryMethodProcessorTest with KSP
This test did not pass as is in KSP, more specifically, turns out
TypeAdapter tries to parse the response type as a Pojo when it is has a
TypeElement, which is always true for KSP. We had a test that returned
int from a multi-column query, which silently passed in JavaC but failed
in KSP as we tried to find a pojo adapter for it.

Making type adapter store check for primitive explicitly fixes the issue
BUT we need an audit of usages of XType.typeElement to ensure
they are not meant as "is this a class or primitive" as that logic does
not hold in KSP anymore. Luckily, we can make it work as we know if it
is a primitive from the TypeName but regardless, it is probably not the
best way to decide during processing.

Bug: 184894249
Bug: 160322705
Test: QueryMethodProcessorTest

Change-Id: I27b2d37045fc5660bfd56c7ecb7e4d84c406974a