Fix a memory leak caused by multiple threads concurrently accessing the same collection
[Root cause]
In Camera2CameraImpl, the methods attachUseCases() and detachUseCases() accept an input UseCase collection, and post tasks to the camera thread to manipulate the collection. However, the collection set by CameraUseCaseAdapter may be changed in the main thread. This will lead to inconsistent UseCase to be attached or detached in the camera thread. Then UseCaseAttachState in Camera2CameraImpl will contain the SessionConfig of the unbound UseCase and the SessionConfig will never be released.

[Solution]
Defensively copy the input collection to prevent other threads from changing it.

Relnote: "Fixed a memory leak caused by multiple threads concurrently accessing the same collection. The memory leak may cause the Activity or Fragment to be retained by a Preview instance."

Bug: 184711452
Test: manually test
Change-Id: I7b4b832be4117cd37f21659555b0ea99e9e56e09