MediaRouter: Fix the button animation
MediaRouteButton shows disconnected state even when
it's connected if it is temporarily disabled.

We could reproduce it by
 button.setEnabled(false)
 button.setEnabled(true)
or
 // select other drawable than "connected"
 mRemoteIndicator.selectDrawable(1);
 // select the "connected" drawble
 mRemoteIndicator.selectDrawable(0);

The reason why it happens is that DrawableContainer#selectDrawable
resets the newly selected drawable, which is AnimationDrawable in our case.

For "connected" state, we should show the last frame of the animation drawable
but we didn't when the drawable state changes while the connection state remains.

I moved the logic for animation into MediaRouteButton#drawableStateChanged and
made sure the exact drawable we want to show is selected.

Bug: 182089242
Test: manually
Change-Id: I00ebf5ce1b21004145aa59bd35044daa9ccb688b