Ignore any events if TapType.CANCEL was received.
The current implementation considered a TapType.CANCEL as a possible
starting point for a tap. However, we should never react to a
TapType.CANCEL, as this means that system has detected a gesture that
is handled by the system, e.g., a long press on the watch face.

Not doing so can lead to the system and the watch face both triggering
an action on the same gesture.

Bug: 184898527
Test: ./gradle :wear:wear-watchface:test and manually tested tap events on the watch face
Change-Id: I178dd93d3e028b5db9b3eb67b26d3f0a22a4a49d