Fix sharedElement transitions for ViewGroups
When doing shared element transitions on ViewGroups with transition
group set to false, we should still allow the ViewGroup to be part of
the transition. This means that the ViewGroup will transition and the
children will transition based on their own parameters (i.e. it is
specifically part of the shared element transition).

RelNote: "When doing a shared element transition on a ViewGroup, if the
ViewGroup has transitionGroup set to false, it will now properly
transition."
Test: testNestedSharedElementViewNonMatching
Bug: 180312148

Change-Id: I99675eac030325415789be0762aa666355f27dd8